LPS-121804 Move logic
