+++ modules/sdk/gradle-plugins/build.gradle
@@ -60,7 +60,6 @@ dependencies {
 	compile group: "com.liferay", name: "com.liferay.gradle.plugins.wsdl.builder", version: "3.0.2"
 	compile group: "com.liferay", name: "com.liferay.gradle.plugins.xsd.builder", version: "1.0.10"
 	compile group: "com.liferay", name: "com.liferay.gradle.util", version: "1.0.44"
-	compile group: "com.pswidersk", name: "python-gradle-plugin", version: "1.2.1"
 
 	compileOnly fileTree(builtBy: [rootProject.tasks.getByName("extractGradleApi" + gradleVersion.replace(".", ""))], dir: new File(rootProject.buildDir, "gradle-${gradleVersion}"), excludes: ["lib/plugins/biz.aQute.bndlib-4.0.0.jar"])
 }
